Make sure to register default preference values
This patch fixes the issue that glyph names are not preserved when
importing UFO files.

UFO.glyphsFileFormat uses the key 'ImportKeepGlyphsNames' when reading a
UFO file to determine whether to convert the glyphs names in the font to
their nice names as defined in GlyphData.xml. Registering default values
seems to be done in the initializer of GSMenu - it should happen when
MainMenu.nib is loaded, which is one of the things we won't do in this
headless version. So instead, this patch manually mimics what Glyphs
does on launch.