Eliminate flaky executable relocation
Stumbled upon another idea to pretend as if we're launched from the
inside of the bundle without relying on what we've used up until now -
to swizzle + [NSBundle mainBundle] and that worked. Note that still the
CFBundleGetMainBundle() counterpart in Core Foundation still remains
untouched and returns the original bundle instance but I'll leave it as
it is for now – patching C function should require another Mach-O magic.