Add facility to fire up specific installation of Glyphs
Now tries to launch a specific version of Glyphs if the environment
variable 'GLYPHSAPP_PATH' is available. The value must be one of the
existing bundle paths of Glyphs in the Applications folder.
The currently installed versions and paths of Glyphs can be listed by
passing '--list-versions' to the command.

Plus, I realised that the recent changes that stopped relocating the
executable breaks the compatibility against the older version of Glyphs.
So I reverted that code in case the given bundle needs such a workaround.